refactor: 함수 파일 이동 및 분리
